feat(inquiries): full back-and-front integration for inquiries and responsive UI
This commit includes:
- Merge latest backend version
- Create InquiryForm component with title, description, and property selector
- Add client-side validation and error alerts
- Create AppointmentComponent (UI only)
- Build InquiryPanel for streamlined inquiry navigation
- Update form formatting and file structure for consistency
- Create ManagerPanel with redirects to specific panels for better component management
- Fetch and render user’s turns in User Profile
- Display inquiry status badges
- Add data table showing all inquiries with filters: All / General / Property / Status

Refs: #33, #35, #36

+++ backend/ms-properties/src/main/java/pi/ms_properties/domain/Property.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.annotation.JsonManagedReference;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.*;
 
@@ -43,13 +44,13 @@ public class Property {
     private Float coveredArea;
 
     @Column(name = "price", nullable = false)
-    private Float price;
+    private BigDecimal price;
 
     @Column(name = "show_price", nullable = false)
     private Boolean showPrice;
 
     @Column(name = "expenses", nullable = true)
-    private Float expenses;
+    private BigDecimal expenses;
 
     @Column(name = "description", nullable = false)
     private String description;
